🔨 refactor: alias 적용
