+++ src/hooks/useInterval/useInterval.ts
@@ -1,5 +1,6 @@
 import { useEffect, useRef, useState } from 'react';
-import { Fn } from '../../types';
+
+type Fn = () => void;
 
 /**
  * 지정된 시간 간격만큼 반복적으로 콜백 함수를 호출하는 훅
