--- src/types/number.ts
@@ -1,6 +1,0 @@
-export type PositiveInteger<T extends number> = `${T}` extends
-  | `-${string}`
-  | `${string}.${string}`
-  | '0'
-  ? never
-  : T;
