+++ src/utils/index.ts
@@ -1,5 +1,6 @@
 export { delayExecution, type CancelToken } from './delayExecution';
 export { throttle } from './throttle';
+export { debounce } from './debounce';
 export { isClient, hasNavigator } from './isClient';
 export { isTouchDevice } from './isTouchDevice';
 export { validators, MatchError } from './validators';
